added more code.
If you guys have time, could you read through this and see if it makes sense? What I have done so far:

1.create a grid.
2. place rooms within the grid.
3. check for overlap.

I added a test case at the bottom, it prints out but I can't make sense of it? Check the link at the top, that is what was aiming to achieve on a smaller scale.
This was my goal for this commit:
https://vazgriz.com/wp-content/uploads/2019/11/dungeon1.png